Add Nginx vhost and remove many recipes. YAGNI.

+++ site-cookbooks/lab/recipes/default.rb
@@ -30,4 +30,13 @@ end
 
 execute "clean-up-bento-scripts" do
   command "rm /home/vagrant/*.sh"
+  not_if { Dir["/home/vagrant/*.sh"].empty? }
 end
+
+include_recipe "nginx::source"
+
+cookbook_file "/etc/nginx/sites-available/querycasts-lab" do
+  source "nginx_vhost"
+end
+
+nginx_site "querycasts-lab"
